Reshuffle portal template parsing (import) code to clean out PortalController
