+++ DNN Platform/Library/DotNetNuke.Library.csproj
@@ -228,6 +228,12 @@
     <Compile Include="Entities\Portals\PortalAliasExtensions.cs" />
     <Compile Include="Entities\Portals\PortalSettingUpdatedEventArgs.cs" />
     <Compile Include="Entities\Portals\PortalTemplateEventArgs.cs" />
+    <Compile Include="Entities\Portals\Templates\IPortalTemplateController.cs" />
+    <Compile Include="Entities\Portals\Templates\IPortalTemplateInfo.cs" />
+    <Compile Include="Entities\Portals\Templates\PortalTemplate.cs" />
+    <Compile Include="Entities\Portals\Templates\PortalTemplateController.cs" />
+    <Compile Include="Entities\Portals\Templates\PortalTemplateImporter.cs" />
+    <Compile Include="Entities\Portals\Templates\PortalTemplateInfo.cs" />
     <Compile Include="Entities\Users\Exceptions\InvalidPasswordException.cs" />
     <Compile Include="Entities\Users\Exceptions\InvalidUserRegisterException.cs" />
     <Compile Include="Entities\Users\UpdateUserEventArgs.cs" />
@@ -489,8 +495,8 @@
     <Compile Include="Obsolete\PortalAliasController.cs" />
     <Compile Include="Obsolete\PortalController.cs" />
     <Compile Include="Entities\Portals\IPortalAliasController.cs" />
-    <Compile Include="Entities\Portals\Internal\IPortalTemplateIO.cs" />
-    <Compile Include="Entities\Portals\Internal\PortalTemplateIO.cs" />
+    <Compile Include="Entities\Portals\Templates\IPortalTemplateIO.cs" />
+    <Compile Include="Entities\Portals\Templates\PortalTemplateIO.cs" />
     <Compile Include="Obsolete\RoleController.cs" />
     <Compile Include="Obsolete\TabController.cs" />
     <Compile Include="Obsolete\TestablePortalController.cs" />
