+++ DNN Platform/Library/Entities/Portals/Templates/IPortalTemplateInfo.cs
@@ -1,0 +1,12 @@
+﻿namespace DotNetNuke.Entities.Portals.Templates
+{
+    public interface IPortalTemplateInfo
+    {
+        string ResourceFilePath { get; }
+        string Name { get; }
+        string CultureCode { get; }
+        string TemplateFilePath { get; }
+        string LanguageFilePath { get; }
+        string Description { get; }
+    }
+}
